requirements.txt
apache-airflow==2.9.3 
great-expectations==0.18.12 
pandas==2.2.3 
sqlalchemy==1.4.52 
pyarrow==17.0.0 
jinja2==3.1.0